+++ src/tasks/mod.rs
@@ -50,11 +50,27 @@ pub fn env() -> Exports {
 }
 
 pub fn all() {
+    for t in serial_tasks() {
+        t.sync_then_update();
+    }
+
     // let's run GitHub Release tasks in serial,
     // to not exacerbate rate limiting,
     // but in parallel with everything else
     let ghr_handle = thread::spawn(|| {
         for t in ghr_tasks() {
+            t.sync_then_update();
+        }
+    });
+
+    let golang_handle = thread::spawn(|| {
+        for t in golang_tasks() {
+            t.sync_then_update();
+        }
+    });
+
+    let nodejs_handle = thread::spawn(|| {
+        for t in nodejs_tasks() {
             t.sync_then_update();
         }
     });
@@ -70,6 +86,8 @@ pub fn all() {
     }
 
     ghr_handle.join().unwrap();
+    golang_handle.join().unwrap();
+    nodejs_handle.join().unwrap();
     rust_handle.join().unwrap();
 }
 
@@ -90,29 +108,47 @@ fn ghr_tasks() -> Vec<Task> {
     ]
 }
 
+fn golang_tasks() -> Vec<Task> {
+    vec![
+      golang::task(),
+      goget::task(), // must be after golang
+    ]
+}
+
+fn nodejs_tasks() -> Vec<Task> {
+    vec![
+        nodejs::task(),
+        npm::task(), // must be after nodejs
+    ]
+}
+
 fn rust_tasks() -> Vec<Task> {
     vec![rustup::task(), rustc::task(), rust::task()]
 }
 
+// these tasks should not be run concurrently with others
+fn serial_tasks() -> Vec<Task> {
+    vec![
+        dotfiles::task(), // must be before "config" tasks
+        vim::task(),      // config, causes vim to take over the terminal
+    ]
+}
+
 fn tasks() -> Vec<Task> {
     vec![
-        dotfiles::task(),  // must be before "config" tasks
         alacritty::task(), // config
         atom::task(),
+        #[cfg(not(windows))]
         bash::task(), // config
         git::task(),
-        golang::task(),
-        goget::task(), // must be after golang
         hyper::task(), // config
         #[cfg(target_os = "macos")]
         macos::task(),
-        nodejs::task(),
-        npm::task(),    // must be after nodejs
         psql::task(),   // config
         ssh::task(),    // config
         tmux::task(),   // config
-        vim::task(),    // config
         vscode::task(), // config
+        #[cfg(not(windows))]
         zsh::task(),    // config
         #[cfg(windows)]
         windows::task(),